Move user generating utils to separate module
